Remove spring boot devtools, spring data r2dbc, mysql connector j and r2dbc mysql dependency - Add spring mongodb reactive and keycloak admin client dependency - Add ValidatorException and KeycloakUserNotFoundException

+++ consumption/src/main/java/me/kirenai/re/consumption/domain/model/dto/GetNourishmentTypePercentageResponse.java
@@ -5,19 +5,19 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import lombok.Builder;
-import me.kirenai.re.consumption.util.enums.NourishmentTypeEnum;
+import me.kirenai.re.consumption.domain.enums.NourishmentTypeEnum;
 
 @Builder
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonDeserialize(as = GetNourishmentTypePercentageResponse.class)
 public record GetNourishmentTypePercentageResponse(
         NourishmentTypeEnum nourishmentType,
-        Integer percentage
+        Double percentage
 ) implements GetNourishmentTypeResponse {
 
     @JsonCreator
     public GetNourishmentTypePercentageResponse(@JsonProperty("nourishmentType") NourishmentTypeEnum nourishmentType,
-                                                @JsonProperty("percentage") Integer percentage) {
+                                                @JsonProperty("percentage") Double percentage) {
         this.nourishmentType = NourishmentTypeEnum.PERCENTAGE;
         this.percentage = percentage;
     }
